Update values to reflect new chassis

+++ Odometry.h
@@ -39,7 +39,7 @@ private:
     static const uint8_t ENCODE_A_2;
     static const uint8_t ENCODE_B_1;
     static const uint8_t ENCODE_B_2;
-    static const int GEAR_RATIO;
+    static const double GEAR_RATIO;
     static const int ENCODER_RESOLUTION;
     static const double WHEEL_RADIUS;
     static const double WHEEL_CIRCUMFERENCE; // in millimeters
